OM-41532: XL API: Properly returning no entities when requesting members of a group with 0 members.
Intent
======
The API for group member resolution should only return entities that are actually in the group.

Background
==========
In the API call to get ServiceEntityApiDTOs for members of a group, an empty group was returning all entities. This is because the repositoryApi call to search entities treated an empty input as "retrieve all". The fix is to avoid calling the repositoryApi when the group member list is empty.

Testing
=======
- Added new unit test for the getMembersByGroupUuid method.
- Verified that clicking on an empty group in the UI returns no entities (via the network tab in developer tools).

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/27915/