OM-40676: Create ActionScript component
RB-27220
Background
In XL will implement the ActionScript functionality as a probe, instead of directly from the platform as in classic

New Function
In this review we create a stub of a probe for ActionScript with a single account property, "name". This probe will (eventually) be automatically created, and so the name will be internally generated.

For now it shows up in the Orchestrator category (see screen shot).