OM-40540: XL Component startup sends default properties config to clustermgr; RB-27834
Background
Each XL Component has a set of external configuration properties which customize its operation. These external configuration properties have been defined in a single file ('factoryInstalledComponents.yml') for all component types. This file is read by ClusterMgr on startup and sets the default configuration for each component type, storing those in the key/value store.
At startup time a component requests its external configuration from the clustermgr based on a component type and component instance id. The configuration returned is the combination of the component instance and component type default properties (instance property wins). Let's call this operation 'request-configuration'

Problem
Using a single file for all configuration is awkward. To either add a new Component or update the external configuration properties for a single component requires editing 'factoryInstalledComponents.yml', building ClusterMgr, and re-running it. Only then will the new (or updated) component be able to run with the updated configuration properties.

Solution
We (a) remove 'factoryInstalledComponents.yml' and all the processing related to it. (b) instead, when a component begins execution, prior to 'request-configuration', the component reads a local 'component_default.properties' file and sends those properties to clustermgr as the New Default Properties for that component type. ClusterMgr saves those component type defaults to the key/value store. The component then proceeds to 'request-configuration', receiving the defaults just uploaded along with the instance properties overlaid.

In other words, every time the component starts the defaults for that component-type are updated.

Note that the way that mediation components are built there is currently only one 'components_default.properties' shared by all probes. I don't believe that will cause an issue - the probe customization is accomplished by 'probe properties' which is completely orthogonal to the component configuration properties here.

Also fixed: It was a mistake early on to declare that the internal-only ClusterMgrRestApi implements the IClusterService interface. IClusterService should strictly define the service interface available to the external-api ClusterController. The internal-only ClusterMgrRestApi should be completely independent of the IClusterService definition.  In fact, in the future the ClusterMgrRestApi should be replaced by a gRPC service.