OM-41734: Exception thrown in ClusterMgr startup if there are not components defined RB-28037
Background
Clustermgr used to use an internal file 'factoryInstalledComponents.yml' to set up default configurations for all component types. This created a component-type list a side-effect. This component-type list was used to "launch" all known components - really just a placeholder for a future when ClusterMgr would talk directly to docker.

Bug
Now that 'factoryInstalledComponents.yml' is no longer used, the very first time that Clustermgr comes up the list of known component types is empty. There was an error check when retrieving the list - if the list was empty that would have been considered a problem loading; but no longer.

Also, there were a couple of properties (e.g. dbSchemaName) with spurious quotes

Fix
The check for empty list of known component types is no longer needed.
Removed the extra quotes