OM-41891 : Revert SettingMapperTest to previous version. RB:28170
SettingsMapperTest relies on local timezone. Since it was failing in a
different timezone, a fix was made in this commit:
26bce4bee6fb966e6a3228214c91cace3800a0dd. But it hasn't fixed the
problem. The right fix is to specify the timezone in the unit test. For
now, we will revert the code to the previous version as it works in the
US timezone and almost everyone(including the build servers) are in the
US timezone. The right fix can be made at a later date.