OM-37241:fix plan actions check boxes are selectable on result page (rb28279)
Problem:

Plan actions should not be selectable, since they are not actionable. But in XL, they are selectable. See link(https://vmturbo.atlassian.net/secure/attachment/88424/88424_2018-07-20_1302.png).

Investigation:

In Classic, all the plan actions will "SUCCEEDED" state, but in XL they are "PENDING_ACCEPT".

Proposed fix:

Align with Classic, set the plan actions' state to  "SUCCEEDED".

Note:

In XL, we could possibly fix it in AO or API, but after investigation, it seems fixing in AO requires lots changes. So this implementation choose fixing it in API component.