OM-40589: XL AO: Implementing action stat roll-ups.
Intent
======
Aggregate action stats by hour, day, and month. This will allow us to query long time ranges with reasonable performance.

Unlike stats and costs, we do the roll-ups at the application level and manually insert the rolled-up rows into the next table in the chain. The primary motivation is to allow us to deal with zero-values better. See "Zero-Value Handling" below.

Flow
====
**LiveActionsStatistician** calls **ActionStatsRollupScheduler**, which schedules an **ActionStatRollup** for every (from table, to table, time range, mgmt unit subgroup) tuple that's ready for rollup. The time range in the tuple is a single unit in the "to table" - an hour when going from latest to hourly, a day from hourly to daily, and so on.

**ActionStatRollup** calls the **ActionStatTable.Reader** of each "from" table (latest, hourly, daily) to roll up the relevant records, and calls the **ActionStatTable.Writer** of each "to" table (hourly, daily, monthly) to insert the rolled-up records into the next time range.

Implementation
==============
- **V1_5__create_action_stats.sql** - the new tables. I'm not adding a separate migration to avoid the extra overhead, since no one uses the action stats tables right now.
- **ActionStatsRollupScheduler** - main scheduling logic, and logic to avoid re-running the same rollups.
- **ActionStatsTable** - interfaces/common objects to provide a common way to access the various tables.
- **BaseActionStatTableReader** - common logic to find which time ranges are ready to be rolled up, and to query the records required for a particular **ActionStatsRollup**.
- **BaseActionStatTableWriter** - common logic to write a rolled up record to a table.
- **RolledUpStatCalculator** - the actual rollup logic to aggregate a list of database records into a single set of numbers (min, max, avg).

Zero-Value Handling
===================
Problem: A particular (mgmt unit subgroup, action group) pair may only have actions in a few snapshots in a time range, and we should optimize for that case. (reminder: an action group is a particular combination of action type, mode, category, and state)

For example, suppose there are 6 action plans in an hour, 5 of them have 0 actions of group 1 for Cluster 4, and 1 of them has 6 actions. When rolling these stats up to the hourly table we want 0 to be the min, 6 to be the max, and 1 ((5 * 0 + 6 * 1) / 6)) to be the average. However, it would be very wasteful to put in the 0-value rows in the database for every (mgmt unit subgroup, action group) pair that DOESN'T have actions in an action plan - especially as the number of action group criteria increases.

Our approach for zero-values is to keep a "snapshot" table along with each "stat" table, and record every snapshot time (i.e. a time that has some actions) in the snapshot table. For any (time, mgmt unit subgroup, action group) tuple where the time is in the snapshot table, we know that a missing entry in the stat table means a value of 0.

Wiki for Action Historical Stats: https://vmturbo.atlassian.net/wiki/spaces/Home/pages/817758213/Action+Historical+Stats

Testing
=======
- Unit tests for all newly added classes.
- System testing - Tested together with query-side
    - Started XL and added vsphere-dc7
    - Verified that actions get saved
    - Verified that action counts are aligned with the actual action list returned.

Reviewed By: garyzeng, psoohoo
https://rbcommons.com/s/VMTurbo/r/27732/