OM-41565: XL AO: Respecting Retention Periods for Action Stats
INTENT
======
Remove stats outside the retention periods specified in global settings. Also, take these settings into account when calculating time frames.

BACKGROUND
==========
We need to trim old records from the various stat tables to avoid ever-growing tables. In action stats we do this via simple delete commands at the application level. Since the volume of data is expected to be lower, we don't need to have the partition logic that we have in history. If we need more efficient/complex procedures in the future we can cross that bridge when we get to it :)

IMPLEMENTATION
==============
Common:
- **RetentionPeriodFetcher** - responsible for getting retention periods from the group component. Caches retention periods locally to avoid overhead of many calls when a batch of queries come in.
- **TimeFrameCalculator** - uses RetentionPeriodFetcher to get the current retention periods, and calculates time frames according to those. For example, if we shorten "hourly" stats to 12 hours, then any query beyond 12 hours ago should look in the "daily" table.
- Moved **TimeFrameCalculator** to components.common instead of components.api.

Action Orchestrator:
- **ActionStatCleanupScheduler** - responsible for scheduling cleanups every X minutes. Obtains the latest retention periods using a RetentionPeriodFetcher, and asks the stat tables to remove records outside the retention range.
- **ActionStatTable** - new methods to calculate the right "trim time", and to remove all entries longer than "trim time" ago (see: **BaseActionStatTableWriter**)

MISCELLANEOUS
=============
- I removed an unnecessary dependency in a pom file, which caused various artifacts - most notably junit - to be unavailable in other poms. Importing junit directly.

TESTING
=======
- Unit tests.
- Integration test to verify that rows actually get deleted.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/28441/