OM-42280: XL Setup: Query an API endpoint instead of an API/UI asset.
Intent
======
The setup XL page should re-direct to the API when the API is ready to serve the UI. (note - technically we should wait until the entire system is up, but that's outside the scope of this fix, since this setup page is going away with K8S).

Background
==========
The way we currently detect if the API is available is by trying to load an image from the UI's assets. However, due to a recent change the UI's assets no longer have deterministic names - each build uses a different hash suffix.

Solution
========
Instead of querying for an image, we query the "/api/v2/checkInit" endpoint, which does not require authentication. If the call succeeds, we know the "external-facing" API servlet is reachable, so we can try to load the UI.

Testing
=======
- Tested the solution with nginx up and API down. We use the same monitor.html page to detect when the API comes back up, so if it works in nginx it should work in setup.
    - Brought up nginx.
    - Navigated to localhost. See the setup page. Timer is ticking.
    - Start API component, and watch the setup page. After ~30 seconds it redirects to the login screen.

Reviewed By: michael.borodiansky, psoohoo
https://rbcommons.com/s/VMTurbo/r/28502/