OM-42463: XL AO: Properly set "Management Unit Type" for mgmt unit subgroups, and use it when querying subgroups.
Problem
=========
The "All Actions" widget was not showing any data when scoped to clusters.

Background
==========
The UI sent the query with the "ON PREM" environment type, but we were saving cluster stats with an "unset" environment type. This change saves all cluster management unit subgroups with the "ON PREM" evironment type, since all clusters are implicitly on prem.

Also, saving the type of management unit in the table. This is mainly for debugging purposes for now - but the row was already there and we just weren't populating it.

Testing
=======
- All unit tests pass. Added unit test to make sure cluster MU Subgroup is retrievable with or without env type set.
- System testing - verified the widget works.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/28549/