OM-36835: XL API: Mapping Storage Cluster to a GroupApiDTO with the "StorageCluster" class name.
Problem
=======
The health widget for a storage cluster was displaying physical machine stats instead of storage stats. It turned out this was because the widget uses the cluster's class name to determine what entity types to look for, and the class name for all clusters was "Cluster".

Solution
========
In GroupMapper (API Component), set the class name of storage clusters to "StorageCluster" instead of "Cluster."

Testing
=======
- Updated unit tests pass.
- System test - brought up XL, verified "Health" widget now shows storages for storage cluster.

Reviewed By: ioannis.kassios
https://rbcommons.com/s/VMTurbo/r/28629/