OM-41693 : Create a reverseIndex from entityId->ClusterId. RB:28326
The breadcrumb in the UI displays the cluster an entity belongs to(if it
exists). To support this feature, we maintain a in-memory mapping from
hostId/storageId -> clusterId in the group component. If the entity is
not a host/storage, we have to find the host that the entity is
connected to by querying the repository component and then from the
hostId we can resolve the clusterId.

Also refactored the Uploading of Groups, Settings and Policies which was
done as one RPC call per TargetId to be be uploaded in a single RPC call
in a streaming fashion(client streaming).