OM-42761 : Use ClusterDisplayName for searching cluster instead of ClusterName. RB:28780
When trying to create groups of VMs, the search filter "Cluster name"
equals "$foo" doesn't work. This is because we are trying to search for
name(which is internal to the group component). We should search by the
display name as it is the one visible to the user in the UI.