OM-42804 : Handle duplicate cluster names. RB:28876
Clusters discovered by different targets can have the same
name(constraint name). A discovered cluster is uniquely identified by
the 3-tuple {constraint_name, entity_type, discovered_by_target_id}.
We correctly store this in the db and display them in the UI. But in
the Entity->Cluster Mapping, we store only one mapping which will lead
to mapping a wrong cluster to an entity). Another issue is that, in
GroupComponent, when we send the Discovered Policies, we specify only
the clusterName in the .proto for the discovered policy. Due to this, in
the group component, we can in-correctly associate the Discovered
policies with the wrong cluster when there are duplicate cluster names.
Fix is to use the 3-tuple for the cluster and group names in TP and
Group component.