OM-41519: XL AO: Add environment type to actions.
(note - the changes look intimidating, but it's mainly because I moved ActionDTOUtil)

INTENT
======
Actions should be filter-able by environment type. Also, clones in the projected topology should have a distinct origin, as well as an environment type and type-specific info.

BACKGROUND
==========
We currently have no reliable way of telling when an action is an ON_PREM or CLOUD action, and we need to be able to do it to support separate ON_PREM and CLOUD queries.

This change adds "environment type" to the ActionEntity message that describes an entity involved in an action. It also introduces an **ActionEnvironmentType** enum, which, given an **Action**, can be ON_PREM, CLOUD, or ON_PREM_AND_CLOUD. We calculate the action's environment type by looking at the types of entities involved in the action. This approach gives us the flexibility of being able to deal with hybrid actions as well.

MAIN CHANGES
============
Protobuf
========
- **ActionDTO.proto** - Add an environment type option to **ActionQueryFilter** and **ActionEntity**.
- **TopologyDTO.proto** - Add a **AnalysisOrigin**, used by entities provisioned by the market (i.e. clones).
- **ActionEnvironmentType** - utility to figure out the environment type from the entities involved in an action.
- Move **ActionDTOUtil**, **ActionEnvironmentType** and **UnsupportedActionException** to com.vmturbo.common.protobuf.action.

Action Orchestrator
====================
- **QueryFilter** - Apply environment time in **ActionQueryFilter**.
- **GlobalActionAggregator** - Using the environment type to determine how to process each action. Note - we're not really handling hybrid actions properly here, because we don't expect them in the "realtime" market right now.

Market
======
- **Analysis** - Set the environment type, origin, and type-specific info of entities in the projected topology.
- **TopologyConverter** - Return a Map<Long, ProjectedTopologyEntity> instead of a list.
- **ActionInterpreter** - Set the environment type of **ActionEntity** objects. It now works with the projected topology instead of a separate oid -> entity type map.

Api
===
- **ActionSpecMapper** - Set the environment type in the action query, if necessary.

MISCELLANEOUS
=============
- Added a "platform" profile - it's basically the development profile without probes.

TESTING
======
- Full build, with additional unit tests.
- Started up XL + a mixed-env-type environment. Verified that the actions tab in HYBRID shows all actions, in CLOUD shows only actions relating to cloud, and ON PREM shows only on prem actions.

Reviewed By: psoohoo, nitya.vyas, thiru
https://rbcommons.com/s/VMTurbo/r/28614/