OM-43371: Added a comprehensive test suite for the Bisection class.
This class, that implements the bisection root-finding method, had a
single test that was enough to make the code appear covered in test
coverage reports, but failed to uncover multiple cases where the method
returned erroneous results and/or failed completely.

This commit adds a comprehensive test suite that covers both usual input
and various edge cases, but leaves the test cases that the current
implementation fails commented out, so that they will stand out.

A subsequent commit will fix the implementation and uncomment all test
cases.

Summary of changes:

* Added ~100 new test cases.
* Added minimum comments required by Turbonomic coding standards and
  then some more. (The existing code didn't meet the requirements)
* Refactored the existing test as one case of a larger parametrised test
  for valid input and unique roots, but otherwise kept it.
* Added a few more digits to a floating point constant that represented
  a repeating decimal in the existing test. That was done to tighten the
  maximum divergence between expected and actual values to 1e-10.