OM-43371: Reimplemented Bisection.solve to fix 2 bugs and improve performance.
There were two bugs in the old implementation:

1. It didn't check for the possibility that the midpoint of the interval
   would happen to be a root. If you were unlucky enough for that to
   happen, it would skip the solution and converge to some arbitrary
   value.
2. The closer the root was to 0, the more iterations it took to converge
   to the root. Although this is arguably a performance problem, it
   becomes a bug if you consider the fact that the function used to
   accept a 'maxIterations' argument and gave up after that many
   iterations.

   Since the client had no way to know the correct number of iterations
   (Integer.MAX_VALUE wasn't always enough either), the method could
   give up too soon producing no value at all.

The new implementation is loosely based on a Julia post
(https://www.shapeoperator.com/2014/02/22/bisecting-floats/) and returns
the closest representable double-precision floating-point value to the
root in no more than 64 iterations.

The above allowed me to simplify the API removing the epsilon,
errorFunction and maxIterations parameters as well as the possibility of
an exception.

Although I didn't write and run proper benchmarks, the fact that the new
implementation takes the same time to pass 105 tests that the old one
needed to pass 52 is an indication that performance has been improved.

Another indication is that the new implementation takes less than a
millisecond to complete in all cases whilst the old one could take up to
multiple seconds to complete or fail if the root was near zero.

While reimplementing the method I also:

* Improved documentation.
* Changed parameter names to use domain (mathematics) terminology.
* Updated clients to make use of the simplified API. That involved
  removing no-longer-used variables and constants and stale comments as
  well as removing a try-catch block and rewrapping lines to take
  advantage of the additional space left from lowering the indentation
  level.
* Added two utility functions that implement an order-preserving mapping
  from doubles to longs and back with accompanying documentation. Tests
  for them will be added in a subsequent commit.