OM-42967: XL AO: Don't retrieve all historical actions when getting action views from LiveActionStore.
Issue
=====
The AO was slowing/crashing (with OOM) in the XL instance managing the lab when a user was navigating the UI.

Background
==========
The action orchestrator was fine when no one was using the UI, but the memory spiked when a user logged in. The various actions-related widgets were taking forever to load, and failing to load in some cases.

The issue was that as part of any action-related query (counts, or the actual actions) we were querying the entire action history table - even if the query wasn't for historical actions. There were ~50k rows in the action history table, and the code was pulling all those rows into memory for every query.

Solution
========
We only really need to query the history table when we want actions by date. This change removes the historical query for "current" action views, and removes the relevant method from the Action History interface to avoid the likelihood of doing this again.

Future Work
===========
I filed OM 43247 for further improvements to the action history interface. Now that we have committed to an approach for action history stats we can make improvements to what we do with "actual" actions.

Testing
=======
- Build.

Reviewed By: emanuele.maccherani, garyzeng
https://rbcommons.com/s/VMTurbo/r/29070/