OM-43371: Gave utility methods in Bisection class package access and tested them.
These methods implement an order-preserving mapping from doubles to
longs and its inverse and are needed by the bisection implementation.

Both specific input-output pairs and abstract properties are tested.
The initial plan was to include more test cases for the abstract
properties (perhaps taken from a random number generator with a known
seed), but the testing framework appears to become sluggish after a
certain number of test cases and thus only a bare minimum (~170) was
included.