OM-42806: XL API: Ensuring string property filter regexes are surrounded with "^" and "$".
Intent
======
Property queries ("name", but also others) should implicitly match the whole property value, and not a subset.

Background
==============
This is to make the behaviour consistent with classic. Also, it makes sense to use a "full" match because the values of properties are always single lines, so we're not matching against a large block of text.

I considered two approaches:
1) Add the "^" and "$" to every query in the API component. This would give us more flexibility down the line, since we could change certain searches to use "contains" while others use exact matching.
2) Add the "^" and "$" in the repository. This has the advantage that we don't need to track down every use of the API to make sure they're consistent.

I ended up going with approach 1) for the extra flexibility, and also because the "contains" behaviour is fairly standard. For example, the implementation of search in the Topology Processor uses a "contains"-type filter, and the MySQL regexp evaluator uses a "contains"-type search. If we do 1) properly then no matter what backend we store data in the search results will be consistent.

This change also adds a migration for all existing groups - adding a "^.\*" prefix and a ".\*$" suffix.

Testing
=======
- Unit testing.
- Brought up XL, tested via Search and also via dynamic group creation.
   - To verify the topology processor side works correctly, created a group that would have contained entities before this change, but contains no entities after, created a setting policy for it (changing automation level), and verified the policy did not affec VMs after broadcast. Then I changed the group to use a ".\*" suffix, re-broadcasted, and verified the policy DID affect the VMs (no actions).

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/29153/