OM-40947:Update movable, cloneable, and suspendable properties based on action capability (rb28819)
Intent:
======
Editor to update following properties based on probes' action capabilities.
1. movable:
a. When entity type is VM:
i. if provider is PM and action type is MOVE, set movable to true unless capability is set to NOT_SUPPORTED.
ii. if provider is Storage and action type is CHANGE, set movable to true unless capability is set to NOT_SUPPORTED.
iii. set movable to false for all other providers.
b. For all other entity types, set movable to true unless capability is set to NOT_SUPPORTED.
2. cloneable:
cloneable is set to false, if there is no probe that can support the provision action
(all the probes/target for that entity have not supported for provision).
3. suspendable:
suspendable is set to false, same as cloneable but for suspend action.

Special case when DTO is discovered by more than one probe.
===========================================================
a. If all probes don't support an action for an entity type, disable the action
for that entity (with same type).
b. If at least one probe doesn't have NOT_SUPPORT on an action for an entity type,
enable the action for that entity (with same type), except the action is disabled by user.
c. If at least one probe doesn't have action capabilities, handle the same ways as b.

(cherry picked from commit d785912)