OM-43060 : Remove dump histogram hook during OOM of an xl java component. RB:29308
At RBC, we noticed that on jvm OOM(for market and vcenter probe), the dump_historgram
was triggered and it was not able to dump any histogram and eventually timed out.
After this, the container was stuck and was not restarted. Disabling
the dump histogram script to prevent any interference from this
script during OOM for the automatic restart of the container.