OM-43920: Fixed a number of comment issues in M2. #3
They were mostly reported by IntelliJ IDEA.

In particular:

* Corrected spelling, grammar and punctuation mistakes.
* Fixed Javadoc syntax errors.
* Made sure Javadoc complied with Turbonomic coding standards.
* Updated some stale comments.

When confronted with Javadoc tags with an empty description (illegal and
in violation of Turbonomic standards) I tried to fill in the
description, when I could deduce it reasonably fast, but removed the tag
completely when not. That was to end up with conforming code and to
concentrate on the rest of the IDE warnings.

This covers the following packages:

* pricefunction
* topology
* translators
* utilities

Plus one small comment fix from the ede package that I missed in the
previous commit.