OM-43498: XL AO 7.12 - Disable Cluster Action Aggregator
Issue
=====
In 7.12.0 the cluster aggregator is way too slow on large topologies. It makes supply chain requests to get information about every cluster, and that took up to 2 hours on an appliance in the lab.

Solution
========
For 7.12.0 the solution is simply to disable to cluster aggregator. We will work on a more proper solution for future releases.

Testing
=======
- Build

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/29399/