OM-44060: XL API: Set "currentEntity" for Activate and Deactivate actions.
Issue
=====
The UI expects "currentEntity" to be set for activate and deactivate (as well as other) actions. We don't currently set it in the ActionApiDTO, which leads to missing explanations for start and stop actions in the UI.

Solution
========
Set the "currentEntity" to be the same as the "targetEntity". The more proper fix would be to change the UI to handle missing current entities, but this is a high-priority issue and the server-side fix is significantly simpler.

Testing
======
- Brought up XL and verified that "SUSPEND" actions now have explanations.
- Built API component, with tests.