OM-43498: XL AO: Refactoring Action Target Selection + Action Support Resolution to improve consistency and performance.
Problem
=======
Action processing and action queries were very slow on large topologies, even if resources (memory, CPU) were not constrained.

Background
==========
There were two existing problems: the performance of ActionTargetByProbeCategoryResolver and the design trio of ActionSupportResolver/ActionTargetSelector/ActionTargetResolver.

ActionTargetByProbeCategoryResolver was being called for every action (by ActionTargetSelector), and was making 3 RPC calls every time - to get probes, targets, and probe capabilities. This meant that for environments with lots of actions action processing was very slow - up to 15 minutes for 60k actions!

Design-wise, ActionSupportResolver used ActionTargetSelector to figure out which target an action should execute on, and resolved the support level for that target. However, underneath, ActionTargetSelector was using ActionTargetResolver to determine the target, and ActionTargetSelector had a different - less accurate - implementation of support checking, because it had to determine which targets were "eligible" to execute an action. This meant duplicate functionality, and potential errors!

Solution
========
- Remove ActionTargetByProbeCategoryResolver, ActionCapabilitiesStore, and ActionSupportResolver.
- Introduce **ProbeCapabilityCache** which has three responsibilities:
    1) Keep a local copy of the capability-related information for each probe.
    2) Given an action, the primary entity, and a probe ID, determine the support level using the cached capabilities.
    3) Given a probe ID, return the probe category.
- Refactor the **ActionTargetSelector** to:
    1) Determine the primary entity of an action - or the primary entities of bulk actions.
    2) Look up the targets that discovered the primary entities.
    3) Use the **ProbeCapabilityCache** to find the support levels for each target.
    4) If multiple targets support the action, use the list of probe priorities to determine which target to pick (get the probe category from the **ProbeCapabilityCache**.
    5) Return the support level and the target ID to the caller.
- Every place that used to use the ActionSupportResolver now uses the ActionTargetSelector directly - most notably the part of the LiveActionStore population code that assigns support levels to each action.
    - Note - we now run the support level calculation outside the action lock, and before constructing the Action objects, so we don't need to have a copy constructor for Actions.

We fully refresh the **ProbeCapabilityCache** every time we process an action plan to be safe. The cache also listens for target and probe updates, so when action execution comes in we will get the most recent data from it.

Testing
======
- Manual testing - brought up the system and got an action plan to display in the UI.
   - Attempted to manually execute the action (it failed in VC, but everything up to then worked fine!)
- Unit testing.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/29555/