OM-43407 - Upgrade arangodb to 3.3.22 version. RB:29678
The existing Arangodb version(3.1.21) is very old. We observed crashes
due to some internal FloationPoint issues in the arangodb engine. Also
when the DB was crashing due to OOM kill or was being abruptly stopped
during operational activieis like upgrade etc, it was leading to DB
corruption. According to arangodb forums, these issues have been fixed
in the 3.3.x versions. So upgrading to 3.3.22 version.

At the time of this commit, the latest stable arangodb version is 3.4.0
which has RocksDB as the default storage engine. Since a newer storage
engine requires more thorough testing, we are not upgrading to 3.4.x
yet.