OM-43715 - Increase nginx read timeout while downloading diags. RB:29600
The diag download takes a lot of time if there are lot of logs or if
some component is slow to dump its state. Due to the small read
timeout setting in nginx, the slower diag downloads were failing. I
have increased the read timeout in nginx to 2 hours for the diag
download request.

Also made changes to ActionOrchestrator(AO) locking so that the diags
download are not blocked when there are new set of actions from the
market to AO.

There are some places, where we are making inter-component calls while
dumping diags. We should remove all such dependencies.Diag dumps should only
dump the data that the component owns. For now, I have removed the call
from API to get the entities count as we can already get this from the
topology dumped by repository or from the TP diags.