OM-42560: XL AO: Sending Topology Info as part of action plan, using topology creation time for action stats.
INTENT
======
The snapshot times for "latest" action stats should be aligned with the snapshot times for "latest" stats in history.

BACKGROUND
==========
The action orchestrator currently uses clock time when saving action stats. This means that the stat snapshot times for action stats are different from those in history, and widgets that combine the two (e.g. Workload Improvements) look very strange. The solution is for both the action orchestrator and history (as well as other components like cost!) to use the topology creation time when saving stats for the topology.

I figured that once we add the topology creation time it would be just as good - and more flexible for the future - to send the entire TopologyInfo of the topology the analysis was run on.

However, I ran into a problem supporting the Buy RI action plan - it doesn't run on a particular topology! I didn't want to create a "fake" topology info with only the topology context ID set.

Suggested refactoring:
- Instead of an enum indicating the action plan type, use a oneof, with additional information for each action plan type. The "market" info will contain the source topology info, and the "buy RI" info will only contain the topology context ID.
- Marking the other fields as deprecated. We will continue setting them for now, but remove them after a bit (maybe this is overkill - we don't save the action plan anywhere).

IMPLEMENTATION
==============
- Key changes are in ActionDTO.proto
    - The rest of the changes are basically refactoring to support the new format.
    - Note: I'm not fully handling the old fields because we're not saving the action plans themselves anywhere, and it's not a tragedy if we fail to process a few action plans after an upgrade.

TESTING
======
- Full build
- Brought up XL, verified action plan received by action orchestrator.

Reviewed By: psoohoo, garyzeng
https://rbcommons.com/s/VMTurbo/r/28652/