OM-44860: XL History: Adding metrics and aggregation information to diags. Upgrading jOOQ/flyway.
Intent
======
Make stats-related debugging from diagnostics easier by including aggregation-related information and history prometheus metrics.

Background
==========
Right now history has no special diags, and it's pretty much impossible to tell from the logs whether aggregation is running smoothly, how many "SOFT" errors there are in BasedbIO and so on. I'm adding some metrics to BasedbIO, and adding a HistoryDiagnostics class to capture information about the past day of aggregations, as well as the latest aggregations for each table type (e.g. vm, pm, and so on).

Also, I noticed that our jOOQ and flyway versions for history were way behind the rest of the system. This was a relic from when we shared the schema with opsmgr, so I updated jOOQ and flyway to be consistent. 3.1.0 is so old it's not even listed as a release on jOOQ's github! (3.2.0 is from 2013)

Testing
======
- Built, brought up history. Did broadcast.
- Downloaded diags via the REST endpoint and manually. Saw that the metadata, aggregation performance, and prometheus metrics were there.

Reviewed By: karthikt, patrick.soohoo
https://rbcommons.com/s/VMTurbo/r/30350/