OM-38961: XL TP: Remove invalid relationships when constructing stitching context, and ignore erroneous errors during analysis
Intent
======
There are cases where the probes send invalid entity graphs to the topology processor. For example, they may say an entity is buying a commodity from an entity that doesn't exist. The Topology Processor should handle those cases.

Background
==========
In this specific case, the App Dynamics probe is sending applications that buy commodities from a provider that doesn't exist. The probe will be fixed as part of a separate bug, but this inconsistency crashes the Topology Pipeline at the stitching stage, and prevents the global broadcast from happening. This is no bueno. In this change I'm changing the current handling of invalid cross-object references in the TopologyStitchingGraph (throw exceptions) to remove edges that don't connect two nodes. For example, any commodities bought where the provider is not a discovered entity get removed. This is the least 'extreme' of the possible ways to handle this problem, and it could cause inconsistencies further down the line.

To allow us to treat these "invalid" entities differently in the market, and see them in the diags, I'm adding a "pipeline errors" field to TopologyEntityDTO. We will populate various error codes into the appropriate field in the EntityPipelineErrors message (for now it's just stitching errors). I'm using a single integer with bits set to indicate various problems, mainly to conserve space in large environments.

Implementation
==============
Protobuf
========
- TopologyEntityDTO - adding EntityPipelineErrors message to hold the various errors encountered for this entity during the topology pipeline. Stitching errors are one of the possible error types (the only one for now).
- Add **StitchingErrors** utility class, which can be initialized directly or from a TopologyEntityDTO, and can be used to check for the presence of varioius stitching-related error codes in the EntityPipelineErrors.

Stitching
=========
- Each **StitchingEntity** now has a **StitchingErrors** object (initially empty). When entities are merged, the **StitchingErrors** of the "from" entity gets added to the merge information of the "onto" entity. Note - we don't use the stitching errors for identification - just the targetId and the entityId.

Topology Processor
==================
- **TopologyStitchingGraph::addStitchingData** now records the errors in the relationships of an entity. It does two things:
    - Removes the erroneous relationships from the underlying EntityDTO.Builder
    - Records the erroneous relationships to debug logs, Prometheus metrics, and the TopologyStitchingEntity's **StitchingErrors** collection.
- **BroadcastStage** now returns a "warning" status if any entities being broadcast have pipeline errors, and prints the number of erroneous entities.
- **ScanDiscoveredSettingPoliciesStage** no longer constructs the TopologyEntityDTOs. We already have a ConstructTopologyFromStitchingContextStage, and it's a lot clearer to use it.

Repository
==========
- Adding **EntityPipelineErrorsRepoDTO** to capture the **EntityPipelineErrors**. This way when we load diags from a customer we can see the errors encountered by the entity.

Market
======
- If an entity had stitching errors during the topology pipeline, mark it as non-controllable.

Testing
=======
- Existing tests pass.
- Some additional tests.
- Loaded and broadcasted the erroneous BofA diags from the bug. There were ~750 entities with errors. Verified the error code was in the repository.

Reviewed By: dblinn, emanuele_maccherani
https://rbcommons.com/s/VMTurbo/r/30454/